t move filter tests
