+++ approvaltests-util-tests/src/test/java/com/spun/util/filters/FilterTest.java
@@ -1,4 +1,4 @@
-package com.spun.util.filters.tests;
+package com.spun.util.filters;
 
 import org.lambda.query.Query;
